requirements.txt
# OE Stack
openeye-toolkits
openeye-floe==0.2.166

# Dev Stack
invoke
pytest
